remove unnecessary escapes from polyline
The polyline format expects "raw" backslashes, and interprets escaped
backslashes as simply occuring twice. This will result in incorrect
linestrings when decoded with official and other third-party decoders.

This changeset removes the extra escaping, allowing the returned string
to be used with other APIs. This was resolved in liblwgeom[0] in 2017.

[0]: https://trac.osgeo.org/postgis/ticket/3713